能够运行，目前还有bug

+++ Assets/Scripts/Systems/SpawnPlayerServerSystem.cs
@@ -63,7 +63,7 @@ namespace FootStone.Kitchen
                 var id = networkServerSystem.RegisterEntity(-1, (ushort) EntityType.Character, spawnPlayer.PlayerId, e);
 
                 var interpolatedState = EntityManager.GetComponentData<CharacterInterpolatedState>(e);
-                interpolatedState.MaterialId = spawnPlayer.PlayerId % 4;
+                interpolatedState.MaterialId =  (byte) (spawnPlayer.PlayerId % 4);
                 EntityManager.SetComponentData(e, interpolatedState);
 
                 EntityManager.SetComponentData(e, new ReplicatedEntityData
